// part9/patientor/server/src/types.ts
export interface Diagnose {
	code: string;
	name: string;
	latin?: string;
}

export enum Gender {
	Male = "male",
	Female = "female",
	Other = "other"
}

export interface Patient {
	id: string;
	name: string;
	dateOfBirth: string;
	ssn: string;
	gender: Gender;
	occupation: string;
}


export type NonSensitivePatientData = Omit<Patient, 'ssn'>;

export type NewPatientEntry = Omit<Patient, 'id'>;